Drop panel toolbox
Since 339e3a1ce8ba48827cc3290711639907937b8584 when in edit mode, the
panel could be configured by clicking anywhere on the panel. This makes
the toolbox to enter panel configuration redundant. It's impossible to
click it without getting the blue highlight for the whole panel and
being clear we're going through a new path.

As a concept having something in the panel whilst a user is trying to
set the sizes of things in the panel is fundemantally wrong.

The context menu is unaffected.